[🛠️project] deploy 파일 수정 (#18)
* 🛠️project : deploy 파일 수정

* 🛠️project : deploy 파일 수정

* 🛠️project : deploy 파일 수정

* 🛠️project : deploy 파일 수정

* 🛠️project : deploy 파일 수정

* 🛠️project : deploy 파일 수정

* 🛠️project : deploy 파일 수정

* 🛠️project : deploy 파일 수정

* 🛠️project : deploy 파일 수정

* 🛠️project : deploy 파일 수정

* 🛠️project : deploy 파일 수정

* 🛠️project : deploy 파일 수정

* 🛠️project : deploy 파일 수정

* 🛠️project : deploy 파일 수정

* 🛠️project : deploy 파일 수정

* 🛠️project : deploy 파일 수정

* ♻️refactor : member테이블 삭제

* ♻️refactor : 시큐리티 설정 수정

* 🛠️project : deploy 파일 수정

* ✅test : test code error 수정

* 🛠️project : 로그 확인

* 🛠️project : 테스트 코드 오류로 인해 테스트 코드 실행 시키지 않도록 함

* 🛠️project : 재실행 하지 않도록 함

* 🛠️project : 시크릿 파일 여부 확인

* 🛠️project : 포트 수정
