+++ api/src/main/java/mutsa/api/controller/auth/AuthController.java
@@ -2,11 +2,13 @@ package mutsa.api.controller.auth;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import mutsa.api.dto.LoginResponseDto;
 import mutsa.api.dto.auth.AccessTokenResponse;
 import mutsa.api.dto.auth.LoginRequest;
+import mutsa.api.dto.user.SignUpUserDto;
 import mutsa.api.service.user.UserService;
 import mutsa.common.exception.BusinessException;
 import mutsa.common.exception.ErrorCode;
@@ -28,6 +30,15 @@ public class AuthController {
     //인증이 필요하지 않은 서비스들
 
     private final UserService userService;
+
+    @PostMapping("/signup")
+    public ResponseEntity<?> signUpUser(@RequestBody @Valid SignUpUserDto signUpDto) {
+        if (!signUpDto.getPassword().equals(signUpDto.getCheckPassword())) {
+            throw new BusinessException(ErrorCode.DIFFERENT_PASSWORD);
+        }
+        userService.signUp(signUpDto);
+        return new ResponseEntity<>("회원가입이 완료되었습니다.", HttpStatus.NO_CONTENT);
+    }
 
     @PostMapping("/login")
     public ResponseEntity<LoginResponseDto> login(HttpServletRequest request, HttpServletResponse response, @Validated @RequestBody LoginRequest loginRequest) throws IOException {
